+++ classlib/java.base/src/main/java/de/mirkosertic/bytecoder/classlib/java/io/TFileOutputStream.java
@@ -60,30 +60,30 @@ public class TFileOutputStream extends OutputStream {
         this.path = null;
     }
 
-    private native long open0(String name, boolean append) throws FileNotFoundException;
+    private native int open0(String name, boolean append) throws FileNotFoundException;
 
-    private native void close0(long handle);
+    private native void close0(int handle);
 
     private void open(final String name, final boolean append) throws FileNotFoundException {
         ((TFileDescriptor) (Object) fd).setHandle(open0(name, append));
     }
 
     public void write(final int b) throws IOException {
-        final long handle = ((TFileDescriptor) (Object) fd).getHandle();
+        final int handle = (int) ((TFileDescriptor) (Object) fd).getHandle();
         writeInt(handle, b);
     }
 
-    public native void writeInt(long handle, int b) throws IOException;
+    public native void writeInt(int handle, int b) throws IOException;
 
-    private native void writeBytes(long handle, byte b[], int off, int len);
+    private native void writeBytes(int handle, byte b[], int off, int len);
 
     public void write(final byte[] b) {
-        final long handle = ((TFileDescriptor) (Object) fd).getHandle();
+        final int handle = (int) ((TFileDescriptor) (Object) fd).getHandle();
         writeBytes(handle, b, 0, b.length);
     }
 
     public void write(final byte[] b, final int off, final int len) {
-        final long handle = ((TFileDescriptor) (Object) fd).getHandle();
+        final int handle = (int) ((TFileDescriptor) (Object) fd).getHandle();
         writeBytes(handle, b, off, len);
     }
 
@@ -91,7 +91,7 @@ public class TFileOutputStream extends OutputStream {
         if (closed) {
             return;
         }
-        final long handle = ((TFileDescriptor) (Object) fd).getHandle();
+        final int handle = (int) ((TFileDescriptor) (Object) fd).getHandle();
         close0(handle);
         closed = true;
     }
